learning: deduplicate initial topic selection
